Conflicts:
	webservice-global/src/main/java/com/sos/joc/classes/JOCResourceImpl.java
	webservice/src/main/java/com/sos/joc/configuration/impl/JocConfigurationResourceImpl.java
	webservice/src/main/java/com/sos/joc/jobchains/impl/JobChainsResourceImpl.java
	webservice/src/main/java/com/sos/joc/jobchains/impl/JobChainsResourcePImpl.java
	webservice/src/main/java/com/sos/joc/jobs/impl/JobsResourceOverviewSnapshotImpl.java
	webservice/src/main/java/com/sos/joc/jobs/impl/JobsResourceOverviewSummaryImpl.java
	webservice/src/main/java/com/sos/joc/jobs/impl/JobsResourcePImpl.java
	webservice/src/main/java/com/sos/joc/jobscheduler/impl/JobSchedulerResourceCommandImpl.java
	webservice/src/main/java/com/sos/joc/locks/impl/LocksResourceImpl.java
	webservice/src/main/java/com/sos/joc/locks/impl/LocksResourcePImpl.java
	webservice/src/main/java/com/sos/joc/orders/impl/OrdersResourceHistoryImpl.java
	webservice/src/main/java/com/sos/joc/orders/impl/OrdersResourceImpl.java
	webservice/src/main/java/com/sos/joc/orders/impl/OrdersResourceOverviewSnapshotImpl.java
	webservice/src/main/java/com/sos/joc/orders/impl/OrdersResourceOverviewSummaryImpl.java
	webservice/src/main/java/com/sos/joc/orders/impl/OrdersResourcePImpl.java
	webservice/src/main/java/com/sos/joc/plan/impl/PlanImpl.java
	webservice/src/main/java/com/sos/joc/processClasses/impl/ProcessClassesResourceImpl.java
	webservice/src/main/java/com/sos/joc/processClasses/impl/ProcessClassesResourcePImpl.java
	webservice/src/main/java/com/sos/joc/schedules/impl/SchedulesResourceImpl.java
	webservice/src/main/java/com/sos/joc/schedules/impl/SchedulesResourcePImpl.java
	webservice/src/main/java/com/sos/joc/tasks/impl/TasksResourceHistoryImpl.java
	webservice/src/main/java/com/sos/joc/tree/impl/TreeResourceImpl.java

+++ webservice/src/main/java/com/sos/joc/configuration/impl/JocConfigurationResourceImpl.java
@@ -144,7 +144,7 @@ public class JocConfigurationResourceImpl extends JOCResourceImpl implements IJo
 			if (configuration.getId() == null || configuration.getId() == 0) {
 				dbItem.setId(null);
 			} else {
-				dbItem.setId(configuration.getId().longValue());
+				dbItem.setId(configuration.getId());
 			}
 
 			/** save item to DB */
